Update based on PR review
