+++ tests/test_opt/test_crfo.py
@@ -143,7 +143,7 @@ def test_sanitised_zero_length_step():
     """Should be able to update with a null step"""
 
     optimiser = CRFOptimiser(etol=1, gtol=1, maxiter=1)
-    optimiser._take_sanitised_step(np.array([]))
+    optimiser._take_step_within_trust_radius(np.array([]))
 
 
 @requires_with_working_xtb_install
